Refactor selection of points for profile.
The responsibility of keeping track of the list of points to profile
(input by the user or selected from a feature/set of features) is now in
profiletool_core.
calculateProfil has been split into updateProfil (and
updateProfilFromFeatures that gets some code that was previously in
SelectLineTool) and plotProfil, so that subsequent calls to plotProfil
can be made to refresh a plot without any other class needing to keep
track of active profiles.
This modification will ease implementation of the following features:
-Updating the profile live while a path is being introduced manually.
-Automatically update the height profile when selection mode changes.